third_party/WebKit/LayoutTests/fast/html/imports/import-custom-element-dup-resolve.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../http/tests/htmlimports/resources/import-custom-element-helper.js"></script>
</head>
<body>
<script>
var Hello = registerTestingCustomElement('x-hello');

var numberOfLinks = 10;
var t1 = async_test('Instantiate custom elements in async imports.');
var t2 = async_test('Resolve custom elements in async imports later.');
var latch = new ImportTestLatch(function() {
    t1.step(function() {
        assert_array_equals(['hello-3', 'hello-4', 'hello-5', 'hello-6', 'hello-p34-1', 'hello-p34-2', 'hello-p56-1', 'hello-p56-2'], Hello.ids.slice(0).sort(), 'All elements are upgraded only once');
        // Remove async links whose loading orders aren't deterministic.
        Hello.ids.splice(Hello.ids.indexOf('hello-4'), 1);
        Hello.ids.splice(Hello.ids.indexOf('hello-6'), 1);
        assert_array_equals(['hello-p34-1', 'hello-3', 'hello-p34-2', 'hello-5', 'hello-p56-1', 'hello-p56-2'], Hello.ids, 'Elements from non async improts are upgraded in order');
        t1.done();
    });
 
    t2.step(function() {
        var Bye = registerTestingCustomElement('x-bye');
        assert_array_equals(['bye-3', 'bye-4', 'bye-5', 'bye-6', 'bye-p34-1', 'bye-p34-2', 'bye-p56-1', 'bye-p56-2'], Bye.ids.slice(0).sort(), 'All elements are upgraded only once');
        // Remove async links whose loading orders are't deterministic.
        Bye.ids.splice(Bye.ids.indexOf('bye-4'), 1);
        Bye.ids.splice(Bye.ids.indexOf('bye-6'), 1);
        assert_array_equals(['bye-p34-1', 'bye-3', 'bye-p34-2', 'bye-5', 'bye-p56-1', 'bye-p56-2'], Bye.ids, 'Elements from non async improts are upgraded in order');
        t2.done();
    });
}, numberOfLinks);
</script>
<link rel="import" href="resources/custom-element-hello-parent-34.html" onload="latch.loaded()">
<link rel="import" href="resources/custom-element-hello-3.html" onload="latch.loaded()">
<link rel="import" href="resources/custom-element-hello-4.html" async onload="latch.loaded()">
<link rel="import" href="resources/custom-element-hello-5.html" onload="latch.loaded()">
<link rel="import" href="resources/custom-element-hello-6.html" async onload="latch.loaded()">
<link rel="import" href="resources/custom-element-hello-parent-56.html" onload="latch.loaded()">
</body>
</html>
